clinic_application/requirements.txt
requests
python-dotenv
pymongo
sv-ttk
